+++ alphadia/utilities/fragment_container.py
@@ -1,10 +1,3 @@
-# native imports
-
-# alphadia imports
-
-# alpha family imports
-
-# third party imports
 import numba as nb
 import numpy as np
